kernel/requirements.txt
Django
django-cors-headers
django-filter
graphene-django
psycopg2
gunicorn
pillow == 10.4.0
requests == 2.32.2
django-redis == 5.4.0
pyjwt == 2.9.0
django-graphql-jwt == 0.4.0
